requirements.txt
environs==9.5.0
python-telegram-bot==11.1.0
redis==4.3.4
requests==2.28.1
